+++ packages/headless/src/features/instant-results/instant-result-analytics-actions.ts
@@ -1,0 +1,32 @@
+import {Result} from '../../api/search/search/result';
+import {
+  partialDocumentInformation,
+  documentIdentifier,
+  validateResultPayload,
+  makeAnalyticsAction,
+  AnalyticsType,
+  InstantResultsSearchAction,
+  InstantResultsClickAction,
+} from '../analytics/analytics-utils';
+
+export const logInstantResultOpen = (
+  result: Result
+): InstantResultsClickAction =>
+  makeAnalyticsAction(
+    'analytics/instantResult/open',
+    AnalyticsType.Click,
+    (client, state) => {
+      validateResultPayload(result);
+      return client.makeDocumentOpen(
+        partialDocumentInformation(result, state),
+        documentIdentifier(result)
+      );
+    }
+  );
+
+export const logInstantResultsSearch = (): InstantResultsSearchAction =>
+  makeAnalyticsAction(
+    'analytics/instantResult/searchboxAsYouType',
+    AnalyticsType.Search,
+    (client) => client.makeSearchboxAsYouType()
+  );